refactor: rename files from Lcp Context to Lcp Presence
