+++ projects/storefrontlib/cms-structure/page/component/services/cms-injector.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { CmsService } from '@spartacus/core';
 import { CmsComponentData, LCP_PRESENCE } from '@spartacus/storefront';
-import { DEFAULT_LCP_PRESENCE } from 'projects/storefrontlib/shared/lcp-context/lcp-context.token';
+import { DEFAULT_LCP_PRESENCE } from 'projects/storefrontlib/shared/lcp-context/lcp-presence.token';
 import { of } from 'rxjs';
 import { CmsComponentsService } from '../../../services/cms-components.service';
 import { CmsInjectorService } from './cms-injector.service';
