+++ projects/storefrontlib/cms-structure/page/component/services/cms-injector.service.ts
@@ -5,7 +5,7 @@
  */
 
 import { Injectable, Injector } from '@angular/core';
-import { provideLcpPresenceForCmsComponent } from '../../../cms-lcp-context/provide-lcp-context-for-cms-component';
+import { provideLcpPresenceForCmsComponent } from '../../../cms-lcp-context/provide-lcp-presence-for-cms-component';
 import { CmsComponentsService } from '../../../services/cms-components.service';
 import { CmsComponentData } from '../../model/cms-component-data';
 import { ComponentDataProvider } from './component-data.provider';
