+++ projects/storefrontlib/shared/lcp-context/lcp-presence.model.ts
